# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Resume-Job Matching System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  details?: string;
}

export interface MatchRequest {
  /** Raw resume text (plain text) */
  resumeText: string;
  /** Job description text */
  jobText: string;
  /** Optional job ID from dataset */
  jobId?: string;
  /** Optional resume ID from dataset */
  resumeId?: string;
}

export type MatchResultDecision =
  (typeof MatchResultDecision)[keyof typeof MatchResultDecision];

export const MatchResultDecision = {
  STRONG_FIT: "STRONG_FIT",
  GOOD_FIT: "GOOD_FIT",
  WEAK_FIT: "WEAK_FIT",
  REJECT: "REJECT",
  INSUFFICIENT_DATA: "INSUFFICIENT_DATA",
} as const;

export interface PenaltyBreakdown {
  domainMismatch: number;
  lowCoreSkillMatch: number;
  experienceGap: number;
  total: number;
  damped: boolean;
}

export type WarningType = (typeof WarningType)[keyof typeof WarningType];

export const WarningType = {
  SEMANTIC_SKILL_CONTRADICTION: "SEMANTIC_SKILL_CONTRADICTION",
  SKILL_SEMANTIC_CONTRADICTION: "SKILL_SEMANTIC_CONTRADICTION",
  EXPERIENCE_SKILL_CONTRADICTION: "EXPERIENCE_SKILL_CONTRADICTION",
  ADVERSARIAL_KEYWORD_STUFFING: "ADVERSARIAL_KEYWORD_STUFFING",
  LOW_CONFIDENCE: "LOW_CONFIDENCE",
} as const;

export interface Warning {
  type: WarningType;
  message: string;
}

export type DomainAlignmentDistribution = { [key: string]: number };

export interface DomainAlignment {
  detectedDomain: string;
  jobDomain: string;
  alignmentScore: number;
  distribution: DomainAlignmentDistribution;
}

export interface MatchResult {
  /** Final score 0-100 */
  finalScore: number;
  decision: MatchResultDecision;
  /** Confidence 0-1 */
  confidence: number;
  semanticScore: number;
  weightedSkillScore: number;
  experienceScore: number;
  baseScore: number;
  penalties: PenaltyBreakdown;
  matchedSkills: string[];
  missingSkills: string[];
  impliedSkills: string[];
  warnings: Warning[];
  explanation: string;
  domainAlignment: DomainAlignment;
  adversarialFlags: string[];
}

export interface CustomResume {
  id: string;
  text: string;
  label?: string;
}

export interface RankRequest {
  jobText: string;
  jobId?: string;
  /** Resume IDs from dataset (optional, uses all if empty) */
  resumeIds?: string[];
  customResumes?: CustomResume[];
}

export interface RankedCandidate {
  rank: number;
  resumeId: string;
  label: string;
  matchResult: MatchResult;
}

export interface PairwiseComparison {
  candidateA: string;
  candidateB: string;
  winner: string;
  scoreDiff: number;
  keyDifferentiator: string;
}

export interface RankResult {
  jobSummary: string;
  ranked: RankedCandidate[];
  top3: RankedCandidate[];
  pairwiseComparisons: PairwiseComparison[];
}

export interface StabilityTestRequest {
  resumeText?: string;
  jobText: string;
  resumeId?: string;
  jobId?: string;
}

export type StabilityTestResultStability =
  (typeof StabilityTestResultStability)[keyof typeof StabilityTestResultStability];

export const StabilityTestResultStability = {
  HIGH: "HIGH",
  MEDIUM: "MEDIUM",
  LOW: "LOW",
} as const;

export interface StabilityTestResult {
  original: MatchResult;
  modified: MatchResult;
  /** Description of the modification applied */
  modification: string;
  scoreVariance: number;
  rankShift: number;
  stability: StabilityTestResultStability;
  analysis: string;
}

export interface Job {
  id: string;
  title: string;
  domain: string;
  description: string;
  requiredSkills: string[];
  preferredSkills: string[];
  experienceRequired: number;
  seniority: string;
}

export interface Resume {
  id: string;
  category: string;
  experienceYears?: number | null;
  source: string;
  syntheticType?: string | null;
  /** First 500 chars of resume text */
  preview: string;
}

export interface SkillEntry {
  canonical: string;
  synonyms: string[];
  implicitPhrases: string[];
  domain: string;
  /** 1.0 = core skill, 0.5 = secondary */
  weight: number;
}

export interface EvaluationCase {
  resumeId: string;
  jobId: string;
  systemDecision: string;
  humanDecision: string;
  systemScore: number;
  humanScore: number;
  agree: boolean;
  failureMode?: string | null;
}

export interface EvaluationMetrics {
  accuracy: number;
  precision: number;
  recall: number;
  falsePositives: number;
  falseNegatives: number;
  totalCases: number;
  agreementScore: number;
  rankCorrelation: number;
  cases: EvaluationCase[];
}

export type ReliabilitySummaryRankingStability =
  (typeof ReliabilitySummaryRankingStability)[keyof typeof ReliabilitySummaryRankingStability];

export const ReliabilitySummaryRankingStability = {
  HIGH: "HIGH",
  MEDIUM: "MEDIUM",
  LOW: "LOW",
} as const;

export interface CostModel {
  falsePositiveCost: string;
  falseNegativeCost: string;
  prioritization: string;
}

export interface ReliabilitySummary {
  agreementScore: number;
  precision: number;
  rankingStability: ReliabilitySummaryRankingStability;
  keyFailureModes: string[];
  decisionPhilosophy: string;
  systemBoundaries: string[];
  costModel: CostModel;
}