# lib/api-zod/src/generated/types/customResume.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Resume-Job Matching System API
 * OpenAPI spec version: 0.1.0
 */

export interface CustomResume {
  id: string;
  text: string;
  label?: string;
}