# lib/api-zod/src/generated/types/domainAlignment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Resume-Job Matching System API
 * OpenAPI spec version: 0.1.0
 */
import type { DomainAlignmentDistribution } from "./domainAlignmentDistribution";

export interface DomainAlignment {
  detectedDomain: string;
  jobDomain: string;
  alignmentScore: number;
  distribution: DomainAlignmentDistribution;
}