# lib/api-zod/src/generated/types/matchRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Resume-Job Matching System API
 * OpenAPI spec version: 0.1.0
 */

export interface MatchRequest {
  /** Raw resume text (plain text) */
  resumeText: string;
  /** Job description text */
  jobText: string;
  /** Optional job ID from dataset */
  jobId?: string;
  /** Optional resume ID from dataset */
  resumeId?: string;
}