# lib/api-zod/src/generated/types/penaltyBreakdown.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Resume-Job Matching System API
 * OpenAPI spec version: 0.1.0
 */

export interface PenaltyBreakdown {
  domainMismatch: number;
  lowCoreSkillMatch: number;
  experienceGap: number;
  total: number;
  damped: boolean;
}